recovery op 4

+++ libespm/include/libespm/BrowserInfo.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <cstdint>
+#include <cstddef>
 
 namespace espm {
 
@@ -8,13 +9,13 @@ class CombineBrowser;
 struct BrowserInfo
 {
   BrowserInfo() = default;
-  BrowserInfo(const CombineBrowser* parent_, uint8_t fileIdx_);
+  BrowserInfo(const CombineBrowser* parent_, size_t fileIdx_);
 
   // Returns 0 for empty (default constructed) LookupResult
   uint32_t ToGlobalId(uint32_t rawId) const noexcept;
 
   const CombineBrowser* const parent = nullptr;
-  const uint8_t fileIdx = 0;
+  const size_t fileIdx = 0;
 };
 
 }
