+++ libespm/include/libespm/LookupResult.h
@@ -8,7 +8,7 @@ struct LookupResult : public BrowserInfo
 {
   LookupResult() = default;
   LookupResult(const CombineBrowser* parent_, const RecordHeader* rec_,
-               uint8_t fileIdx_);
+               size_t fileIdx_);
 
   const RecordHeader* const rec = nullptr;
 };
