+++ libespm/src/LookupResult.cpp
@@ -3,7 +3,7 @@
 namespace espm {
 
 LookupResult::LookupResult(const CombineBrowser* parent_,
-                           const RecordHeader* rec_, uint8_t fileIdx_)
+                           const RecordHeader* rec_, size_t fileIdx_)
   : BrowserInfo(parent_, fileIdx_)
   , rec(rec_)
 {
